feat(app): fix type errors
