+++ app/src/lib/utils.ts
@@ -5,18 +5,3 @@ import type {ClassValue} from "clsx";
 export function cn(...inputs: Array<ClassValue>) {
   return twMerge(clsx(inputs))
 }
-
-export function formatDuration(minutes: number): string {
-  const hours = Math.floor(minutes / 60);
-  const mins = minutes % 60;
-
-  if (hours === 0) {
-    return `${mins} min`;
-  }
-
-  if (mins === 0) {
-    return `${hours} hr${hours > 1 ? "s" : ""}`;
-  }
-
-  return `${hours} hr${hours > 1 ? "s" : ""} ${mins} min`;
-}
